requirements.txt
pandas==2.2.2
requests==2.32.3
psycopg[binary]==3.2.2
python-dateutil==2.9.0.post0